third_party/WebKit/LayoutTests/external/wpt/web-animations/interfaces/Animation/idlharness-expected.txt

This is a testharness.js-based test.
PASS Animation interface automated IDL tests 
FAIL Animation interface: existence and properties of interface object assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface object length assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface object name assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: existence and properties of interface prototype object assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: attribute id assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: attribute effect assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: attribute timeline assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: attribute startTime assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: attribute currentTime assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: attribute playbackRate assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: attribute playState assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: attribute ready assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: attribute finished assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: attribute onfinish assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: attribute oncancel assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: operation cancel() assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: operation finish() assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: operation play() assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: operation pause() assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation interface: operation reverse() assert_own_property: self does not have own property "Animation" expected property "Animation" missing
FAIL Animation must be primary interface of new Animation() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Stringification of new Animation() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "id" with the proper type (0) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "effect" with the proper type (1) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "timeline" with the proper type (2) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "startTime" with the proper type (3) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "currentTime" with the proper type (4) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "playbackRate" with the proper type (5) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "playState" with the proper type (6) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "ready" with the proper type (7) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "finished" with the proper type (8) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "onfinish" with the proper type (9) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "oncancel" with the proper type (10) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "cancel" with the proper type (11) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "finish" with the proper type (12) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "play" with the proper type (13) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "pause" with the proper type (14) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
FAIL Animation interface: new Animation() must inherit property "reverse" with the proper type (15) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Animation is not defined"
Harness: the test ran to completion.
